fix: grid mode for images and video content page, shadow not fully displayed when the mouse hovers over the image
In the grid browsing mode of images and on the video content display page, when the mouse hovers over an image, the shadow is not fully displayed and appears truncated at the top.
I identified that the issue was caused by the "inset-y-1" style in the 「ListItemHoverOverlay」 component.
As this is a foundational component, I opted not to modify it directly. Instead, I introduced an "overlayItemClassName" attribute to the 「EntryItemWrapper」 component to address the style issue.