fix: alignment of the title style at the top of the right content page
the type title at the top of the right content page is misaligned across different subscription types.
In the "EntryListHeader" component, an array named "titleStyleBasedView" is defined to configure the padding for various subscription types.
I think this is unnecessary. I attempted to remove it and instead apply a consistent padding value of "pl-7".